+++ src/storage/mod.rs
@@ -1,8 +1,3 @@
-#[cfg(feature = "serde")]
-mod codec_json;
-#[cfg(feature = "prost")]
-mod codec_prost;
-mod codec_string;
 mod use_local_storage;
 mod use_session_storage;
 mod use_storage;