+++ src/utils/codecs/mod.rs
@@ -1,0 +1,4 @@
+mod bin;
+mod string;
+
+pub use string::*;