+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
+mod codecs;
 mod filters;
 mod is;
 mod js_value_from_to_string;
@@ -5,6 +6,7 @@ mod pausable;
 mod signal_filtered;
 mod use_derive_signal;
 
+pub use codecs::*;
 pub use filters::*;
 pub use is::*;
 pub(crate) use js_value_from_to_string::*;